# contrib/google.golang.org/grpc: Reduce allocs in interceptor